cryptohack-writeups/src/requirements.txt
pycryptodome
pwntools>=4.9.0
